Copied contents back from PieChart-fix.jsx to PieChart.jsx
